flask-app/requirements.txt
flask
flask-cors
Pillow
torch
torchvision